Improve test_pickles script
Revert EPIC_train_object_action_correspondence.pkl to 3f31a91992884e2b0f90784e619972c21269e248
since Mike updated that pkl in that commit so the index of the pickle
matches the first column of the CSV in line with the rest of our other
pickles.

Improvements to the testing script mean that we no longer need to
specify the csv paths---we ensure a csv exists for every pickle.